Add early ending feature to stop when balance <= 0.
Truncate the output data after the first month where the end balance
is less than or equal to zero, if this occurs before the input term length.
This prevents interest being accrued on a negative balance.
Note: most loan calculators size the monthly payments to result in an
exactly zero balance at the end of the loan term.